Move non-markup utilities to new yaml utilities module
