+++ tests/unit/test_markup_utils.py
@@ -163,58 +163,3 @@ class TestHtmlRender(TestCase):
         assert '<blockquote>\n<p>NOTE:' in html
         assert 'Tool \'feature\' is <em>beta</em>' in html
         assert '<ul>\n<li>Item1</li>' in html
-
-
-@pytest.fixture
-def tmp_file():
-    try:
-        dir = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
-        tmp_file = os.path.join(dir, 'env.yml')
-        yield tmp_file
-    finally:
-        os.remove(tmp_file)
-
-
-EX_ENV_YAML = """version: 1
-dependencies:
-  python:
-    - somepkg==1.3
-    - otherpkg>=3.0
-  files:
-    - "/usr/bin/oc"
-    - "/usr/lib/libssl.so.1"
-  system:
-    - python3.6-dateutil
-"""
-
-
-def test_get_execution_environment(tmp_file):
-    data = {
-        "version": 1,
-        "dependencies": {
-            "python": [
-                "somepkg==1.3",
-                "otherpkg>=3.0"
-            ],
-            "files": [
-                "/usr/bin/oc",
-                "/usr/lib/libssl.so.1"
-            ],
-            "system": [
-                "python3.6-dateutil"
-            ]
-        }
-    }
-    with open(tmp_file, 'w') as f:
-        f.write(EX_ENV_YAML)
-        f.flush()
-        res = markup_utils.get_execution_environment(tmp_file)
-        assert res == data
-
-
-def test_lint_execution_environment_file(tmp_file):
-    with open(tmp_file, 'w') as f:
-        f.write(EX_ENV_YAML)
-        f.flush()
-        res = list(markup_utils.lint_execution_environment_file(tmp_file))
-        assert len(res) == 1
